+++ src/libraries/simulation_kernel/Entities/Agents/Units/Humans/MIL_Injury_Wound.cpp
@@ -139,7 +139,7 @@ float MIL_Injury_Wound::GetAgentDose() const
 // -----------------------------------------------------------------------------
 float MIL_Injury_Wound::SetLifeExpectancy() const
 {
-    return MIL_MedicalTreatmentType::Find( injuryID_ )->GetLifeExpectancy( injuryCategory_ )*( 1 + 0.1*MIL_Random::rand_ii( -1 , 1 ) );
+    return static_cast< float >( MIL_MedicalTreatmentType::Find( injuryID_ )->GetLifeExpectancy( injuryCategory_ )*( 1 + 0.1*MIL_Random::rand_ii( -1 , 1 ) ) );
 }
 
 // -----------------------------------------------------------------------------
